+++ baybe/utils/dataframe.py
@@ -9,6 +9,7 @@ from typing import TYPE_CHECKING, Any, Literal, TypeVar, overload
 
 import numpy as np
 import pandas as pd
+from typing_extensions import assert_never
 
 from baybe.exceptions import InputDataTypeWarning, SearchSpaceMatchWarning
 from baybe.parameters.base import Parameter
@@ -41,31 +42,51 @@ def to_tensor(*x: _ConvertibleToTensor) -> Tensor | tuple[Tensor, ...]:
 
     Returns:
         The provided inputs represented as tensor(s).
+
+        If possible, returns views of the original data, otherwise creates copies.
+        The latter will be the case if the input data is not contiguous in memory, not
+        of the configured Torch dtype, or uses negative strides.
     """
-    # FIXME This function seems to trigger a problem when some columns in either of
-    #  the dfs have a dtype other than int or float (e.g. object, bool). This can
-    #  weirdly happen, even if all values are numeric, e.g. when a target column is
-    #  looked up from a df in simulation, it can have dtype object even if it's all
-    #  floats. As a simple fix (this seems to be the most reasonable place to take
-    #  care of this) df.values has been changed to df.values.astype(float),
-    #  even though this seems like double casting here.
     import torch
 
-    from baybe.utils.torch import DTypeFloatTorch
+    from baybe.utils.torch import DTypeFloatTorch, torch_to_numpy_dtype_mapping
 
-    out = tuple(
-        torch.tensor(xi, dtype=DTypeFloatTorch)
-        if isinstance(xi, (int, float))
-        else torch.from_numpy(
-            (xi.values if isinstance(xi, (pd.Series, pd.DataFrame)) else xi).astype(
-                DTypeFloatNumpy
-            )
-        ).to(DTypeFloatTorch)
-        for xi in x
-    )
+    numpy_dtype = torch_to_numpy_dtype_mapping[DTypeFloatTorch]
+
+    def _convert(x: _ConvertibleToTensor, /) -> Tensor:
+        match x:
+            case int() | float():
+                return torch.tensor(x, dtype=DTypeFloatTorch)
+            case np.ndarray():
+                # tensors with negative strides are not supported by PyTorch
+                fix_strides = any(s < 0 for s in x.strides)
+                x = x.astype(numpy_dtype, copy=fix_strides)
+                tensor = torch.from_numpy(x)
+            case pd.Series() | pd.DataFrame():
+                # We already coerce to the target dtype during the dataframe-to-numpy
+                # conversion since this step might otherwise return an array of type
+                # `object` in case of mixed column types, which would cause the
+                # subsequent numpy-to-torch conversion to fail. This happens, for
+                # example, when the dataframe contains Boolean and integer columns.
+
+                # tensors with negative strides are not supported by PyTorch
+                fix_strides = any(s < 0 for s in x.to_numpy().strides)
+                array = x.to_numpy(numpy_dtype, copy=fix_strides)
+                tensor = torch.from_numpy(array)
+            case _:
+                assert_never(x)
+
+        # The `contiguous` call brings us closest to getting reproducible
+        # results downstream in the torch ecosystem
+        # https://github.com/meta-pytorch/botorch/issues/3046
+        return tensor.contiguous()
+
+    converted = tuple(_convert(xi) for xi in x)
+
     if len(x) == 1:
-        out = out[0]
-    return out
+        return converted[0]
+
+    return converted
 
 
 def add_fake_measurements(