+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Methods `hshift` and `vshift` to `Transformation` for conveniently performing
   horizontal / vertical shifts
 
+### Changed
+- Dataframe-to-tensor conversion now yields contiguous tensors, improving
+  reproducibility of downstream operations
+
 ### Fixed
 - Random seed not entering simulation when explicitly passed to `simulate_scenarios`